requirements.txt
python-decouple
Django>=4.0
psycopg2-binary
djangorestframework
djangorestframework-simplejwt